requirement.txt
psycopg2
python-dotenv
langchain
langchain_groq
